requirements.txt
numpy==1.21
flask
librosa
tensorflow-cpu
